requirements.txt
discord
pillow
python-chess
cairosvg
deflate
zstandard